requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
djangorestframework==3.15.2
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
filetype==1.2.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pdf2image==1.17.0
pillow==11.1.0
psycopg2==2.9.10
PyPDF2==3.0.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
